requirements.txt
openai
python-dotenv
wikipedia
tiktoken
neo4j
streamlit
# add speech to text
streamlit_extras
streamlit_mic_recorder
sentence_transformers==2.2.2
Pillow
fastapi
PyPDF2
torch==2.0.1
pydantic
uvicorn
sse-starlette
boto3
# atlassian
atlassian-python-api
lxml
pytesseract
# missing from the langchain base image?
langchain-openai
langchain-community==0.0.19
langchain-google-genai
# url bot
beautifulsoup4
chromadb